refactor gasstation config
