+++ gasstation/gasstattion_test.go
@@ -6,170 +6,12 @@
 package gasstation
 
 import (
-	"context"
-	"fmt"
-	"math/big"
 	"testing"
 
 	"github.com/stretchr/testify/require"
-
-	"github.com/iotexproject/iotex-core/action"
-	"github.com/iotexproject/iotex-core/action/protocol"
-	"github.com/iotexproject/iotex-core/action/protocol/account"
-	accountutil "github.com/iotexproject/iotex-core/action/protocol/account/util"
-	"github.com/iotexproject/iotex-core/action/protocol/execution"
-	"github.com/iotexproject/iotex-core/action/protocol/rewarding"
-	"github.com/iotexproject/iotex-core/action/protocol/rolldpos"
-	"github.com/iotexproject/iotex-core/actpool"
-	"github.com/iotexproject/iotex-core/blockchain"
-	"github.com/iotexproject/iotex-core/blockchain/block"
-	"github.com/iotexproject/iotex-core/blockchain/blockdao"
-	"github.com/iotexproject/iotex-core/config"
-	"github.com/iotexproject/iotex-core/db"
-	"github.com/iotexproject/iotex-core/pkg/unit"
-	"github.com/iotexproject/iotex-core/state/factory"
-	"github.com/iotexproject/iotex-core/test/identityset"
-	"github.com/iotexproject/iotex-core/testutil"
 )
 
 func TestNewGasStation(t *testing.T) {
 	require := require.New(t)
-	require.NotNil(NewGasStation(nil, nil, config.Default.API.GasStation))
-}
-func TestSuggestGasPriceForUserAction(t *testing.T) {
-	ctx := context.Background()
-	cfg := config.Default
-	cfg.Genesis.BlockGasLimit = uint64(100000)
-	cfg.Genesis.EnableGravityChainVoting = false
-	registry := protocol.NewRegistry()
-	acc := account.NewProtocol(rewarding.DepositGas)
-	require.NoError(t, acc.Register(registry))
-	rp := rolldpos.NewProtocol(cfg.Genesis.NumCandidateDelegates, cfg.Genesis.NumDelegates, cfg.Genesis.NumSubEpochs)
-	require.NoError(t, rp.Register(registry))
-	factoryCfg := factory.GenerateConfig(cfg.Chain, cfg.Genesis)
-	sf, err := factory.NewFactory(factoryCfg, db.NewMemKVStore(), factory.RegistryOption(registry))
-	require.NoError(t, err)
-	ap, err := actpool.NewActPool(cfg.Genesis, sf, cfg.ActPool)
-	require.NoError(t, err)
-	blkMemDao := blockdao.NewBlockDAOInMemForTest([]blockdao.BlockIndexer{sf})
-	bc := blockchain.NewBlockchain(
-		cfg.Chain,
-		cfg.Genesis,
-		blkMemDao,
-		factory.NewMinter(sf, ap),
-		blockchain.BlockValidatorOption(block.NewValidator(
-			sf,
-			protocol.NewGenericValidator(sf, accountutil.AccountState),
-		)),
-	)
-	ep := execution.NewProtocol(blkMemDao.GetBlockHash, rewarding.DepositGas)
-	require.NoError(t, ep.Register(registry))
-	rewardingProtocol := rewarding.NewProtocol(cfg.Genesis.Rewarding)
-	require.NoError(t, rewardingProtocol.Register(registry))
-	require.NoError(t, bc.Start(ctx))
-	defer func() {
-		require.NoError(t, bc.Stop(ctx))
-	}()
-
-	for i := 0; i < 30; i++ {
-		tsf, err := action.NewTransfer(
-			uint64(i)+1,
-			big.NewInt(100),
-			identityset.Address(27).String(),
-			[]byte{}, uint64(100000),
-			big.NewInt(1).Mul(big.NewInt(int64(i)+10), big.NewInt(unit.Qev)),
-		)
-		require.NoError(t, err)
-
-		bd := &action.EnvelopeBuilder{}
-		elp1 := bd.SetAction(tsf).
-			SetNonce(uint64(i) + 1).
-			SetGasLimit(100000).
-			SetGasPrice(big.NewInt(1).Mul(big.NewInt(int64(i)+10), big.NewInt(unit.Qev))).Build()
-		selp1, err := action.Sign(elp1, identityset.PrivateKey(0))
-		require.NoError(t, err)
-
-		require.NoError(t, ap.Add(context.Background(), selp1))
-
-		blk, err := bc.MintNewBlock(testutil.TimestampNow())
-		require.NoError(t, err)
-		require.Equal(t, 2, len(blk.Actions))
-		require.Equal(t, 2, len(blk.Receipts))
-		var gasConsumed uint64
-		for _, receipt := range blk.Receipts {
-			gasConsumed += receipt.GasConsumed
-		}
-		require.True(t, gasConsumed <= cfg.Genesis.BlockGasLimit)
-		require.NoError(t, bc.CommitBlock(blk))
-	}
-	height := bc.TipHeight()
-	fmt.Printf("Open blockchain pass, height = %d\n", height)
-
-	gs := NewGasStation(bc, blkMemDao, cfg.API.GasStation)
-	require.NotNil(t, gs)
-
-	gp, err := gs.SuggestGasPrice()
-	require.NoError(t, err)
-	// i from 10 to 29,gasprice for 20 to 39,60%*20+20=31
-	require.Equal(t, big.NewInt(1).Mul(big.NewInt(int64(31)), big.NewInt(unit.Qev)).Uint64(), gp)
-}
-
-func TestSuggestGasPriceForSystemAction(t *testing.T) {
-	ctx := context.Background()
-	cfg := config.Default
-	cfg.Genesis.BlockGasLimit = uint64(100000)
-	cfg.Genesis.EnableGravityChainVoting = false
-	registry := protocol.NewRegistry()
-	acc := account.NewProtocol(rewarding.DepositGas)
-	require.NoError(t, acc.Register(registry))
-	rp := rolldpos.NewProtocol(cfg.Genesis.NumCandidateDelegates, cfg.Genesis.NumDelegates, cfg.Genesis.NumSubEpochs)
-	require.NoError(t, rp.Register(registry))
-	factoryCfg := factory.GenerateConfig(cfg.Chain, cfg.Genesis)
-	sf, err := factory.NewFactory(factoryCfg, db.NewMemKVStore(), factory.RegistryOption(registry))
-	require.NoError(t, err)
-	ap, err := actpool.NewActPool(cfg.Genesis, sf, cfg.ActPool)
-	require.NoError(t, err)
-	blkMemDao := blockdao.NewBlockDAOInMemForTest([]blockdao.BlockIndexer{sf})
-	bc := blockchain.NewBlockchain(
-		cfg.Chain,
-		cfg.Genesis,
-		blkMemDao,
-		factory.NewMinter(sf, ap),
-		blockchain.BlockValidatorOption(block.NewValidator(
-			sf,
-			protocol.NewGenericValidator(sf, accountutil.AccountState),
-		)),
-	)
-	ep := execution.NewProtocol(blkMemDao.GetBlockHash, rewarding.DepositGas)
-	require.NoError(t, ep.Register(registry))
-	rewardingProtocol := rewarding.NewProtocol(cfg.Genesis.Rewarding)
-	require.NoError(t, rewardingProtocol.Register(registry))
-	require.NoError(t, bc.Start(ctx))
-	defer func() {
-		require.NoError(t, bc.Stop(ctx))
-	}()
-
-	for i := 0; i < 30; i++ {
-		blk, err := bc.MintNewBlock(testutil.TimestampNow())
-		require.NoError(t, err)
-		require.Equal(t, 1, len(blk.Actions))
-		require.Equal(t, 1, len(blk.Receipts))
-		var gasConsumed uint64
-		for _, receipt := range blk.Receipts {
-			gasConsumed += receipt.GasConsumed
-		}
-		require.True(t, gasConsumed <= cfg.Genesis.BlockGasLimit)
-		require.NoError(t, bc.CommitBlock(blk))
-	}
-	height := bc.TipHeight()
-	fmt.Printf("Open blockchain pass, height = %d\n", height)
-
-	gs := NewGasStation(bc, blkMemDao, cfg.API.GasStation)
-	require.NotNil(t, gs)
-
-	gp, err := gs.SuggestGasPrice()
-	fmt.Println(gp)
-	require.NoError(t, err)
-	// i from 10 to 29,gasprice for 20 to 39,60%*20+20=31
-	require.Equal(t, gs.cfg.DefaultGas, gp)
+	require.NotNil(NewGasStation(nil, nil, DefaultConfig))
 }
